Moved reducers as subfolder to store to have all redux items in same location
