--- imports/ui/store/reducer.js
@@ -1,8 +1,0 @@
-const initialState = {
-  merchants: [],
-  orders: [],
-}
-
-export default function reducer(state = initialState, action) {
-
-}
