+++ imports/ui/store/reducers/merchantsReducer.js
